* tree-scalar-evolution.c (dump_chrecs_stats): Prevent format warning.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/lno-branch@80249 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/tree-scalar-evolution.c
@@ -3237,7 +3237,7 @@ dump_chrecs_stats (FILE *file)
 	   stats_nb_undetermined);
   fprintf (file, "-----------------------------------------\n");
   fprintf (file, "%d\tchrecs in the scev database\n", 
-	   VARRAY_ACTIVE_SIZE (scalar_evolution_info));
+	   (int) VARRAY_ACTIVE_SIZE (scalar_evolution_info));
   fprintf (file, "-----------------------------------------\n");
   fprintf (file, ")\n\n");
 }
